requirements.txt
# Add all modules that are required for the project to run here

Flask==3.0.2
Flask-Login==0.6.3
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1
Pillow==11.1.0
